+++ HISTORY.md
@@ -1,5 +1,12 @@
 # Revision History for `matdb`
 
+## Revision 1.1.5
+- Updated the Manual Group so that it can take an optional extractable
+  flag which disapbles calulations when set to fals.
+- Updated `database/__init__.py' so that the user can input a
+  transformation dictionary that will change the seed files passed in
+  according to a transformation function.
+
 ## Revision 1.1.4
 - Removed assignments to empty dicts from the function definitions as
   mention in Issue [#56](/../../issues/56).